Integrando com o banco de dados

+++ src/main/java/com/chat/chat/repository/SolicitacaoRepository.java
@@ -2,12 +2,9 @@ package com.chat.chat.repository;
 
 import com.chat.chat.model.Solicitacao;
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import java.util.List;
 
 public interface SolicitacaoRepository extends JpaRepository<Solicitacao, Long> {
-
-    List<Solicitacao> findBySupervisorId(Long supervisorId);
-
-    List<Solicitacao> findByPsicologoId(Long psicologoId);
+    List<Solicitacao> findBySupervisorId(Integer supervisorId);
+    List<Solicitacao> findByPsicologoId(Integer psicologoId);
 }
